requirements.txt
# Core Data Processing & Spark
pyspark==3.5.1
pandas>=2.0.0
numpy>=1.24.0

# Kafka & API Interaction
kafka-python==2.0.2
requests>=2.31.0

# Database
pymongo>=4.6.0

# Frontend & UI
streamlit>=1.32.0
Pillow>=10.2.0
PyYAML>=6.0.1
watchdog>=4.0.0  # Recommended for Streamlit hot-reloading